LayoutTests/http/tests/digital-credentials/digital-credential-all-invalid-requests.https-expected.txt

CONSOLE MESSAGE: An error occurred validating the incoming 'org-iso-mdoc' request. The request will be ignored. (The given request could not parse into an ISO 18013-7 (Annex C) request.)
CONSOLE MESSAGE: An error occurred validating the incoming 'org-iso-mdoc' request. The request will be ignored. (The given request could not parse into an ISO 18013-7 (Annex C) request.)
CONSOLE MESSAGE: An error occurred validating the incoming 'org-iso-mdoc' request. The request will be ignored. (The given request could not parse into an ISO 18013-7 (Annex C) request.)
Tests that navigator.credentials.get() rejects with TypeError when all org-iso-mdoc requests fail framework validation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS promise1 rejected promise  with TypeError: No valid credential requests remain after validation.
PASS promise2 rejected promise  with TypeError: No valid credential requests remain after validation.
PASS successfullyParsed is true

TEST COMPLETE
